cli: use define for argument lists in settings.c
settings.c basically consists of property-type structures and *a lot* of
accessors. All these accessors share the same argument list.

It is very repetitive to specify it over and over again. Also, there
are so many arguments that one is compelled to wrap the lines. All
in all it results in a lot of noise that takes the eye from the important
code.

Also, the argument list is expected to change, so we possibly only
have to change one place.