pacrunner: remove failed and pending items from configuration list
If a configuration does not have a path it is because we are still
sending it to pacrunner or because we failed to do so. In both cases,
we have to remove the configuration from the list.

Fixes: 3ad89223d0ea9a772b650842d15583d92cf9a904
(cherry picked from commit fad2cf072143762d9f51c9d79c05f5ec995b099f)