Handle carrier changes by bond devices
This batch changes the way carrier changes are handled by bond
devices. With this patch:
* Pulling cable out of one of the slaves brings slave into disconnected state.
* Reconnecting cable reconnects slave automatically.
* Pulling out cables of all slaves brings down bond interface - removing its IP configuration.
* Reconnecting cable to one of the slaves makes bond interface reistablish its IP configuration.

With this patch bond interface is made to work like any other interface supporting carrier.
This patch has also been tested with bond interface having one ethernet and one wifi device - and it works as expected.

This patch allows one to have bond interface between wifi and ethernet and stil be able to connect to other wifi networks and not have ip config or routing clashes.

Signed-off-by: Nikolay Martynov <mar.kolya@gmail.com>