bluetooth: split _find_network_server() in two functions
Instead of having a complex _find_network_server() function with several
arguments, split it.

Having multiple optional arguments to a find() function is fine,
if all arguments consistently narrow down the search.
For example nmp_cache_lookup_link_full(), where each argument is
optional, and it restricts the search.
For _find_network_server() that was not the case, because setting
"addr" and "device" together would be non-sensical.